requirements.txt
#mtgsdk==1.3.1
python-dotenv==0.21.1
discord==2.1.0
requests==2.28.2